refactor(): display quote in second fragment
