+++ app/src/main/java/fr/idnow/imagecapture/presentation/FirstFragment.kt
@@ -5,14 +5,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
-import fr.idnow.imagecapture.MainActivity
 import fr.idnow.imagecapture.R
 import fr.idnow.imagecapture.databinding.FragmentFirstBinding
-import fr.idnow.imagecapture.presentation.viewmodels.QuoteViewModel
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 /**
  * A simple [Fragment] subclass as the default destination in the navigation.
@@ -25,20 +20,12 @@ class FirstFragment : Fragment() {
     // onDestroyView.
     private val binding get() = _binding!!
 
-    private lateinit var quoteViewModel: QuoteViewModel
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-
-        val appContainer = (activity as MainActivity).appContainer
-
-        quoteViewModel = QuoteViewModel(appContainer.getSingleQuoteUseCase, Dispatchers.IO)
-
         _binding = FragmentFirstBinding.inflate(inflater, container, false)
         return binding.root
-
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -47,20 +34,6 @@ class FirstFragment : Fragment() {
         binding.buttonFirst.setOnClickListener {
             findNavController().navigate(R.id.action_FirstFragment_to_SecondFragment)
         }
-
-        lifecycleScope.launch {
-            quoteViewModel.quoteUiState.collect { uiState ->
-                when (uiState) {
-                    is QuoteViewModel.QuoteUiState.Success -> binding.tvQuote.text =
-                        uiState.quote.text
-
-                    is QuoteViewModel.QuoteUiState.Error -> binding.tvQuote.text =
-                        uiState.exception.localizedMessage
-                }
-            }
-        }
-
-
     }
 
     override fun onDestroyView() {
